+++ src/components/WebComponentProject/WebComponentProject.jsx
@@ -10,7 +10,11 @@ import MobileProject from "../Mobile/MobileProject/MobileProject";
 import Output from "../Editor/Output/Output";
 import { defaultMZCriteria } from "../../utils/DefaultMZCriteria";
 import Sk from "skulpt";
-import { setIsSplitView, setWebComponent } from "../../redux/EditorSlice";
+import {
+  setIsSplitView,
+  setWebComponent,
+  setIsOutputOnly,
+} from "../../redux/EditorSlice";
 import { MOBILE_MEDIA_QUERY } from "../../utils/mediaQueryBreakpoints";
 import {
   codeChangedEvent,
@@ -35,7 +39,9 @@ const WebComponentProject = ({
   const codeRunTriggered = useSelector(
     (state) => state.editor.codeRunTriggered,
   );
+
   const error = useSelector((state) => state.editor.error);
+  const errorDetails = useSelector((state) => state.editor.errorDetails);
   const codeHasBeenRun = useSelector((state) => state.editor.codeHasBeenRun);
   const currentStepPosition = useSelector(
     (state) => state.instructions.currentStepPosition,
@@ -46,6 +52,7 @@ const WebComponentProject = ({
 
   dispatch(setIsSplitView(false));
   dispatch(setWebComponent(true));
+  dispatch(setIsOutputOnly(outputOnly));
 
   useEffect(() => {
     setCodeHasRun(false);
@@ -69,14 +76,14 @@ const WebComponentProject = ({
       const mz_criteria = Sk.sense_hat
         ? Sk.sense_hat.mz_criteria
         : { ...defaultMZCriteria };
-      document.dispatchEvent(
-        runCompletedEvent({
-          isErrorFree: error === "",
-          ...mz_criteria,
-        }),
-      );
+
+      const payload = outputOnly
+        ? { errorDetails }
+        : { isErrorFree: error === "", ...mz_criteria };
+
+      document.dispatchEvent(runCompletedEvent(payload));
     }
-  }, [codeRunTriggered, codeHasRun, error]);
+  }, [codeRunTriggered, codeHasRun, outputOnly, error, errorDetails]);
 
   useEffect(() => {
     document.dispatchEvent(stepChangedEvent(currentStepPosition));